requirements.txt
asn1crypto==1.5.1
boto3==1.42.68
botocore==1.42.68
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.5
cryptography==46.0.5
filelock==3.25.2
idna==3.11
jmespath==1.1.0
numpy==2.4.3
packaging==26.0
pandas==3.0.1
platformdirs==4.9.4
py4j==0.10.9.9
pyarrow==23.0.1
pycparser==3.0
PyJWT==2.12.1
pyOpenSSL==25.3.0
pyspark==4.1.1
python-dateutil==2.9.0.post0
pytz==2026.1.post1
requests==2.32.5
s3transfer==0.16.0
six==1.17.0
snowflake-connector-python==4.3.0
sortedcontainers==2.4.0
tomlkit==0.14.0
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
pyspark
pandas
numpy
snowflake-connector-python
snowflake-sqlalchemy
python-dotenv
